requirements.txt
openai
Flask
schedule
playwright
cos-python-sdk-v5
PyYAML
chinesecalendar
requests